backend/requirements.txt
asgiref==3.10.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dnspython==2.8.0
gunicorn==23.0.0
idna==3.11
mongoengine==0.29.1
oauthlib==3.3.1
packaging==25.0
pycparser==3.0
PyJWT==2.11.0
pymongo==4.15.5
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.3
tzdata==2025.3
urllib3==2.6.3
